RhythmPrism3 updates
Now with descriptions!

There's still a lot of new modules to build, but you'll find.
atten~ attenuates a signal with a MIDI controllable fader and a secondary input. (All faders should be MIDI assignable.)
average~ combines the inputs of up to four signals and averages them.
directory~ loads a directory of samples that can be triggered sequentially or randomly.
envel~ An ADSR envelope generator with adjustments for exponential/linear/logarithmic slopes.
function~ A ASR envelope generator with separate exp/lin/log slopes for attack and release.
launcher~ provides easy buttons to create new objects with unique identifiers. It has eight slots for saving and loading presets for all objects. Will eventually also have a fader to control transition time between states.
midigen~ A four channel signal to MIDI file generator. Likely will need some adjustments before an acceptable final version.
offset~ adds an offset to a signal from a fader and a secondary input.
pulse~ a button that sends a trigger signal.
range~ displays the minimum and maximum of a continuous signal with faders to set these parameters
relabi~ A four channel feedback loop of sine waves with inputs to control global frequency and to reset the phase. Also includes four sawtooth outputs synced to the sine waves -- great for triggers!
seqrand~ A CV sampler/randomizer. An incoming signal is sampled by a trigger and stored as a sequence. Another signal sets a range of sampled values from most recent to the index number. A second trigger grabs a random value from that range outputing the stored CV. A final input biases the randomness to choose values that are more recent or less recent.
shiftreg~ Is based on the modular synth "Turing Machine" which cycles an 8 bit shift register, sending high and low values out the first eight outlets. The ninth outlet outputs a value between 0-10V (0-1 in Pd) that is formed from the binary combination of the shift register values. Every time shiftreg~ is triggered, it cycles the 8 bits with a chance for the first bit to flip. This chance is set by the flip% fader combined with a signal in the second inlet.
temperament~ is a four channel quantizer. Each channel has an inlet for a control signal and an inlet to sample that signal with a trigger. All four channels are quantized to a 1v/o tuning in a chromatic scale (bottom row of buttons) and quarter-tones (top row). There is a portamento fader. Eight different sets of notes (chords/scales) can be saved. These states can be linked to the launcher's save/load state or act independently.
triggate~ Sends a high gate signal when the first inlet's signal is higher than the threshold slider combined with an optional signal in the right inlet.
vco~ An LFO/VCO which can you can toggle between with the upper left button. Takes signal inputs to control v/o, FM, width, and phase reset, along with faders to control these parameters. You can also toggle the oscillator to be positive or positive and negative. Outputs are sine, triangle, sawtooth, and variable width square. All shapes respond to the width control.