Updates to Rhythm Prism
Notable inclusion of cco~ a sample-rate clock controlled oscillator with width and phase control.

Includes a wrapper for the external vstplugin~. You'll need to use deken to download a copy of that library for your platform.

New object, pattern~, allows the creation of a sequence of up to 16 gate values.

Object midigen~ now has a clock inlet on the far right that will set the bpm of the MIDI file.